requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
jinja2
python-dotenv
anthropic
python-multipart
